requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
PyPDF2==2.5.0
python-docx==0.8.11
requests==2.26.0
gunicorn==20.1.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.5.0
opencv-python==4.8.0.76
PyMuPDF>=1.23.0
python-dotenv==1.0.1
google-generativeai==0.3.2
streamlit==1.24.0
nltk==3.8.1
easyocr>=1.6.0
torch>=1.10.0
# Ke